requirements.txt
fastapi~=0.104.0
uvicorn~=0.23.2
python-multipart
pillow~=10.0.1
tensorflow-serving-api~=2.14.1
matplotlib
numpy~=1.24.3
tensorflow~=2.14.0